Loot Distribution
Added parameters for loot distribution: +rarity for suppressors
+settings for which firearms to spawn(WIP)
Started a re-work of the formatting of the Firearms_Items_ files
+New Spas-12 and Law-12 model as well as a new model for the shotgun suppressor
+New model for MAC-10 and Uzi, as well as a new model for the sling attachment for these firearms.
Fixed code in the Firearms_Silencer.lua file, previously the Sandbox settings did not apply at all. Unknown as to why there was a problem at all, but moved the parameters into the function itself and it seems to have fixed the issue.